# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SignageOS - Digital Signage Management Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /screens:
    get:
      operationId: listScreens
      tags: [screens]
      summary: List all screens
      parameters:
        - name: clientId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of screens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Screen"
    post:
      operationId: createScreen
      tags: [screens]
      summary: Register a new screen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreenInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Screen"
  /screens/{id}:
    get:
      operationId: getScreen
      tags: [screens]
      summary: Get a screen by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Screen details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Screen"
    patch:
      operationId: updateScreen
      tags: [screens]
      summary: Update a screen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreenUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Screen"
    delete:
      operationId: deleteScreen
      tags: [screens]
      summary: Delete a screen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /screens/pair:
    post:
      operationId: pairScreen
      tags: [screens]
      summary: Register a new screen via user pairing code (called from TVBox)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PairScreenInput"
      responses:
        "201":
          description: Screen registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairScreenResult"
        "404":
          description: Pairing code not found
...
  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: List all schedules
      parameters:
        - name: screenId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Create a new schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
  /schedules/{id}:
    patch:
      operationId: updateSchedule
      tags: [schedules]
      summary: Update a schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: Delete a schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /schedules/broadcast:
    post:
      operationId: broadcastPlaylist
      tags: [schedules]
      summary: Send a playlist to ALL screens of the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastInput"
      responses:
        "200":
          description: Broadcast applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BroadcastResult"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    Screen:
      type: object
      required: [id, name, clientId, code, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        code:
          type: string
          description: Unique code used by the TV player URL
        location:
          type: ["string", "null"]
        status:
          type: string
          description: "online, offline, unknown"
        lastSeen:
          type: ["string", "null"]
        activePlaylistName:
          type: ["string", "null"]
        createdAt:
          type: string
    ScreenInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        clientId:
          type: integer
        location:
          type: string
    PairScreenInput:
      type: object
      required: [pairingCode, name]
      properties:
        pairingCode:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        location:
          type: string
    PairScreenResult:
      type: object
      required: [id, name, code, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        location:
          type: [string, 'null']
        status:
          type: string
        createdAt:
          type: string
    BroadcastInput:
      type: object
      required: [playlistId]
      properties:
        playlistId:
          type: integer
    BroadcastResult:
      type: object
      required: [count]
      properties:
        count:
          type: integer
    ScreenUpdate:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        clientId:
          type: integer
...
    Schedule:
      type: object
      required: [id, screenId, playlistId, createdAt]
      properties:
        id:
          type: integer
        screenId:
          type: integer
        screenName:
          type: ["string", "null"]
        playlistId:
          type: integer
        playlistName:
          type: ["string", "null"]
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        daysOfWeek:
          type: ["string", "null"]
          description: "Comma-separated day numbers: 0=Sun, 1=Mon, ... 6=Sat"
        active:
          type: boolean
        createdAt:
          type: string
    ScheduleInput:
      type: object
      required: [screenId, playlistId]
      properties:
        screenId:
          type: integer
        playlistId:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        daysOfWeek:
          type: string
        active:
          type: boolean
    ScheduleUpdate:
      type: object
      properties:
        playlistId:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        daysOfWeek:
          type: string
        active:
          type: boolean
...