# lib/api-zod/src/generated/types/broadcastInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SignageOS - Digital Signage Management Platform
 * OpenAPI spec version: 0.1.0
 */

export interface BroadcastInput {
  playlistId: number;
}