# lib/api-zod/src/generated/types/pairScreenResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SignageOS - Digital Signage Management Platform
 * OpenAPI spec version: 0.1.0
 */

export interface PairScreenResult {
  id: number;
  name: string;
  code: string;
  /** @nullable */
  location?: string | null;
  status: string;
  createdAt: string;
}